requirements.txt
streamlit>=1.12.0
pandas>=2.1.0,<2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=3.3.0
nba_api>=1.2.0
requests>=2.31.0
plotly>=6.1.1
kaleido>=1.2.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
shap>=0.42.0
tweepy>=4.16.0
pillow>=9.0.0
tenacity>=8.0.0
praw>=7.7.0
vaderSentiment>=3.3.2
pywin32>=306
